@@ -1,13 +1,19 @@
 import NIOCore
+import NIOConcurrencyHelpers
 import Logging
 
 extension PostgresDatabase {
     public func simpleQuery(_ string: String) -> EventLoopFuture<[PostgresRow]> {
-        var rows: [PostgresRow] = []
-        return simpleQuery(string) { rows.append($0) }.map { rows }
+        let rowsBoxed = NIOLockedValueBox([PostgresRow]())
+        return self.simpleQuery(string) { row in
+            rowsBoxed.withLockedValue {
+                $0.append(row)
+            }
+        }.map { rowsBoxed.withLockedValue { $0 } }
     }
     
-    public func simpleQuery(_ string: String, _ onRow: @escaping (PostgresRow) throws -> ()) -> EventLoopFuture<Void> {
+    @preconcurrency
+    public func simpleQuery(_ string: String, _ onRow: @Sendable @escaping (PostgresRow) throws -> ()) -> EventLoopFuture<Void> {
         self.query(string, onRow: onRow)
     }
 }
